+++ test/fixtures3/modcache/github.com/acme/iface@v1.2.0/inline.go
@@ -1,0 +1,5 @@
+package iface
+
+type InlineExecutor interface { Inline() error }
+
+type ExtendedInlineExecutor interface { InlineExecutor }